# api/openapi.yaml
openapi: 3.1.0
info:
  title: MCPDoll Control Plane API
  version: 0.1.0
  summary: The single contract behind MCPDoll's API, CLI, and console.
  description: |
    This document is the source of truth for every MCPDoll operation. The Go
    server interfaces, the Go client the CLI uses, the TypeScript client the
    console uses, and `docs/api/` are all generated from it.

    Two rules make that more than a slogan:

    - **The CLI is an API client and never touches the database.** Anything the
      CLI can do, an external integrator can do, because the CLI has no
      privileged path.
    - **Every operation declares its surfaces** in `x-mcpdoll-surfaces`, and
      `make parity` fails the build if an operation lacks a CLI command or a UI
      route, or if a CLI or UI binding names an operation that does not exist.

    See `docs/adr/0004-api-first-tri-surface.md`.

    ## What this version covers

    The registry is a reviewable YAML document rather than database rows, and
    the control plane's durable side is not built (see `docs/deferred.md`). So
    the operations here are the ones that are genuinely implemented: reading the
    registry, inspecting and verifying snapshots, and observing a live data
    plane. Registry *mutation* through the API arrives with the database.

  license:
    name: MIT
    # SPDX, so the declaration is machine-readable and matches ./LICENSE rather
    # than only agreeing with it by eye.
    identifier: MIT
servers:
...
paths:
...
  /api/v1/registry:
    get:
      operationId: getRegistry
      summary: Read the resolved registry document
      tags: [registry]
      x-mcpdoll-surfaces:
        cli: mcpdoll registry show
        ui: /registry
      responses:
        "200":
          description: The registry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Registry" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/registry:validate:
    post:
      operationId: validateRegistry
      summary: Validate a registry document without contacting any backend
      description: |
        Structure and internal consistency only: unknown keys, dangling
        references, duplicate prefixes, TTLs that try to widen rather than
        narrow. Offline, so it belongs in a pull-request check.
      tags: [registry]
      x-mcpdoll-surfaces:
        cli: mcpdoll registry validate
        ui: /registry/validate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ValidateRegistryRequest" }
      responses:
        "200":
          description: The document is valid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegistrySummary" }
        "422":
          description: The document is invalid, with every problem listed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/registry/servers/{serverId}:
    get:
      operationId: getServer
      summary: Read one backend
      tags: [registry]
      x-mcpdoll-surfaces:
        cli: mcpdoll registry servers show
        ui: /registry/servers/:serverId
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          description: The backend
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Server" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/schedules/{jobType}:
    get:
      operationId: getSchedule
      summary: Read one piece of timed work
      tags: [schedules]
      x-mcpdoll-surfaces:
        cli: mcpdoll schedules show
        ui: /schedules/:jobType
      parameters:
        - { name: jobType, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateSchedule
      summary: Retune or switch off one piece of timed work
      description: |
        Changing a cadence takes effect without a restart, and re-arms the next
        run from now — otherwise lengthening an interval would appear not to
        have taken until the old one elapsed.
      tags: [schedules]
      x-mcpdoll-surfaces:
        cli: mcpdoll schedules set
        ui: /schedules/:jobType/edit
      parameters:
        - { name: jobType, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateScheduleRequest" }
      responses:
        "200":
          description: The updated schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/schedules/{jobType}:run:
    post:
      operationId: runScheduleNow
      summary: Bring a schedule forward so it runs on the next tick
      description: |
        Does not run the job inline. It moves the next run to now and the
        scheduler picks it up, through the same claim and the same outcome
        recording as any other run — a second execution path would be a second
        way for the job to happen, and the two would drift.
      tags: [schedules]
      x-mcpdoll-surfaces:
        cli: mcpdoll schedules run
        ui: /schedules/:jobType/run
      parameters:
        - { name: jobType, in: path, required: true, schema: { type: string } }
      responses:
        "202":
          description: The schedule, now due
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/snapshots/current:
    get:
      operationId: getCurrentSnapshot
      summary: Read the snapshot the gateway is serving
      tags: [snapshots]
      x-mcpdoll-surfaces:
        cli: mcpdoll snapshot current
        ui: /snapshots
      responses:
        "200":
          description: The snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Snapshot" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/snapshots:inspect:
    post:
      operationId: inspectSnapshot
      summary: Show what a snapshot file contains
      description: |
        Reads a snapshot *without* verifying it, so one signed by a key you do
        not hold can still be inspected — which is exactly when inspection is
        most useful. Use `verifySnapshot` to check the signature.
      tags: [snapshots]
      x-mcpdoll-surfaces:
        cli: mcpdoll snapshot inspect
        ui: /snapshots/inspect
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InspectSnapshotRequest" }
      responses:
        "200":
          description: The snapshot's contents
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Snapshot" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/snapshots:build:
    post:
      operationId: buildSnapshot
      summary: Resolve the registry into a signed snapshot
      description: |
        Discovers every backend the registry names, canonicalizes what they
        publish, and signs the result. Every problem is a build failure: a
        snapshot some data-plane instances would refuse is worse than none.
      tags: [snapshots]
      x-mcpdoll-surfaces:
        cli: mcpdoll snapshot build
        ui: /snapshots/build
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BuildSnapshotRequest" }
      responses:
        "200":
          description: The build report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BuildReport" }
        "422":
          description: The snapshot could not be built
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/snapshots:verify:
    post:
      operationId: verifySnapshot
      summary: Check a snapshot's signature and that it would activate
      description: |
        Both halves matter. A correctly signed snapshot with a dangling
        reference is refused by every data-plane instance.
      tags: [snapshots]
      x-mcpdoll-surfaces:
        cli: mcpdoll snapshot verify
        ui: /snapshots/verify
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifySnapshotRequest" }
      responses:
        "200":
          description: The verification result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerifyReport" }
        "422":
          description: The snapshot failed verification
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/keys:generate:
    post:
      operationId: generateSigningKey
      summary: Mint an Ed25519 snapshot signing keypair
      description: |
        For local development. The private key is written unencrypted; whoever
        holds it can publish configuration to every data-plane instance, so
        production keys must come from your key management system.
      tags: [snapshots]
      x-mcpdoll-surfaces:
        cli: mcpdoll keys generate
        ui: /snapshots/keys
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GenerateSigningKeyRequest" }
      responses:
        "200":
          description: The generated key
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SigningKey" }
...
  /api/v1/gateway/status:
    get:
      operationId: getGatewayStatus
      summary: Report which snapshot a data plane is serving
      tags: [gateway]
      x-mcpdoll-surfaces:
        cli: mcpdoll gateway status
        ui: /gateway
      responses:
        "200":
          description: The gateway's status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GatewayStatus" }
        "503":
          description: The gateway is not ready
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GatewayStatus" }
  /api/v1/gateway/backends:
    get:
      operationId: listBackends
      summary: What the gateway's prober knows about each backend
      description: |
        Every backend's condition, and how far each has drifted from what was
        admitted. Drift is the interesting column: the gateway serves admitted
        definitions and never live backend output, which makes a backend
        changing its catalog a detectable event rather than a silent change to
        what clients see — but only because something looks.

        `blocked_tools` is what a strict backend's drift actually costs. In
        advisory mode it is always zero however far the backend has moved.
      tags: [gateway]
      x-mcpdoll-surfaces:
        cli: mcpdoll gateway backends
        ui: /gateway/backends
      responses:
        "200":
          description: Backend conditions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BackendHealthReport" }
        "502": { $ref: "#/components/responses/Unavailable" }
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Sign in as a person, and receive a session
      description: |
        A local password is a principal: this returns a session whose grants are
        the user's, and every operation is then checked against them. That was
        recorded as blocked on an identity provider and was not — OIDC is a
        second *source* of identity, not a prerequisite for having one
        (ADR 0022).

        The tenant is part of the identity rather than a lookup hint: the same
        email may exist in two tenants and they are different people.

        Every failure answers identically. A caller learns whether they signed
        in and nothing else — not whether the tenant exists, not whether the
        email does, not whether the account is disabled.
      tags: [auth]
      x-mcpdoll-surfaces:
        cli: mcpdoll auth login
        ui: /login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "201":
          description: The session, with its token shown once
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/NoDatabase" }
...
  /api/v1/tenants/{tenantId}:
    delete:
      operationId: deleteTenant
      summary: Delete a tenant and everything it owns
      description: |
        Cascades to the tenant's users, their grants, and their API keys. The
        cascade is the schema's, which is what makes the deletion complete
        without anybody having to remember every table.
      tags: [tenancy]
      x-mcpdoll-surfaces:
        cli: mcpdoll tenants delete
        ui: /tenants/:tenantId/delete
      parameters:
        - name: tenantId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/NoDatabase" }
...
components:
...
  responses:
    NotFound:
      description: The resource does not exist
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    ValidateRegistryRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: |
            The registry document as YAML. Validation runs against what the
            caller sends rather than what is on the server's disk, which is
            what makes this usable as a pull-request check.
...